+++ core/src/main/scala/ai/lum/odinson/state/file/FileState.scala
@@ -1,5 +1,8 @@
-package ai.lum.odinson.state
+package ai.lum.odinson.state.file
 
+import ai.lum.odinson.state.ResultItem
+import ai.lum.odinson.state.State
+import ai.lum.odinson.state.StateFactory
 import com.typesafe.config.Config
 
 class FileState extends State {
